+++ mne/io/fiff/tests/test_raw_fiff.py
@@ -6,7 +6,6 @@
 
 from copy import deepcopy
 from functools import partial
-import itertools as itt
 import os.path as op
 import pickle
 import sys
@@ -26,8 +25,6 @@ from mne import (concatenate_events, find_events, equalize_channels,
                  pick_info)
 from mne.utils import (_TempDir, requires_pandas, object_diff,
                        requires_mne, run_subprocess, run_tests_if_main)
-from mne.io.proc_history import _get_rank_sss
-from mne.io.pick import _picks_by_type
 from mne.annotations import Annotations
 
 testing_path = testing.data_path(download=False)
@@ -200,37 +197,6 @@ def test_copy_append():
     raw_full.append(raw)
     data = raw_full[:, :][0]
     assert_equal(data.shape[1], 2 * raw._data.shape[1])
-
-
-@pytest.mark.slowtest
-@testing.requires_testing_data
-def test_rank_estimation():
-    """Test raw rank estimation."""
-    iter_tests = itt.product(
-        [fif_fname, hp_fif_fname],  # sss
-        ['norm', dict(mag=1e11, grad=1e9, eeg=1e5)]
-    )
-    for fname, scalings in iter_tests:
-        raw = read_raw_fif(fname).crop(0, 4.).load_data()
-        (_, picks_meg), (_, picks_eeg) = _picks_by_type(raw.info,
-                                                        meg_combined=True)
-        n_meg = len(picks_meg)
-        n_eeg = len(picks_eeg)
-
-        if len(raw.info['proc_history']) == 0:
-            expected_rank = n_meg + n_eeg
-        else:
-            expected_rank = _get_rank_sss(raw.info) + n_eeg
-        assert_array_equal(raw.estimate_rank(scalings=scalings), expected_rank)
-        assert_array_equal(raw.estimate_rank(picks=picks_eeg,
-                                             scalings=scalings), n_eeg)
-        if 'sss' in fname:
-            raw.add_proj(compute_proj_raw(raw))
-        raw.apply_proj()
-        n_proj = len(raw.info['projs'])
-        assert_array_equal(raw.estimate_rank(tstart=0, tstop=3.,
-                                             scalings=scalings),
-                           expected_rank - (0 if 'sss' in fname else n_proj))
 
 
 @testing.requires_testing_data
